refactor(frontend): group src by feature, colocate tests
Flatten src/renderer/ — main.ts, App.svelte, styles.css move to src/.
Components regrouped under src/features/{library,playlist,player,scan,paths,toolbar,track}/.
Shared modules (api, state, types, mockBackend, state.test) under src/shared/.
Drop tests/ tree; vitest now collects src/**/*.test.ts via single root config.
index.html relocates to repo root (Vite root = repo root, Tauri Svelte template convention).
