+++ index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./main.ts"></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
