+++ src/main.ts
@@ -1,7 +1,7 @@
 import { mount } from "svelte";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import App from "./App.svelte";
-import { app } from "./state.svelte";
+import { app } from "./shared/state.svelte";
 import "./styles.css";
 
 mount(App, { target: document.getElementById("app")! });
