--- src/renderer/App.svelte
@@ -1,19 +1,0 @@
-<script lang="ts">
-  import Toolbar from "./components/Toolbar.svelte";
-  import NowPlaying from "./components/NowPlaying.svelte";
-  import LibraryPanel from "./components/LibraryPanel.svelte";
-  import PlaylistPanel from "./components/PlaylistPanel.svelte";
-  import PathsOverlay from "./components/PathsOverlay.svelte";
-  import ScanStatusBar from "./components/ScanStatusBar.svelte";
-  import TrackTooltip from "./components/TrackTooltip.svelte";
-</script>
-
-<Toolbar />
-<NowPlaying />
-<main id="content">
-  <LibraryPanel />
-  <PlaylistPanel />
-</main>
-<PathsOverlay />
-<ScanStatusBar />
-<TrackTooltip />
