--- tests/renderer/vitest.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "vitest/config";
-import { svelte } from "@sveltejs/vite-plugin-svelte";
-
-export default defineConfig({
-  plugins: [svelte({ configFile: "../../svelte.config.mjs" })],
-  test: {
-    name: "renderer",
-    environment: "jsdom",
-    include: ["**/*.test.ts"],
-  },
-});
