+++ tsconfig.node.json
@@ -5,7 +5,6 @@
     "types": ["node"]
   },
   "include": [
-    "src/types.ts",
     "commitlint.config.ts",
     "eslint.config.mjs",
     "prettier.config.mjs",
