+++ vite.config.ts
@@ -5,7 +5,6 @@ import { resolve } from "node:path";
 const host = process.env["TAURI_DEV_HOST"];
 
 export default defineConfig({
-  root: resolve(__dirname, "src/renderer"),
   publicDir: false,
   plugins: [svelte({ configFile: resolve(__dirname, "svelte.config.mjs") })],
   clearScreen: false,
