+++ vitest.config.ts
@@ -1,8 +1,12 @@
 import { defineConfig } from "vitest/config";
+import { svelte } from "@sveltejs/vite-plugin-svelte";
 
 export default defineConfig({
+  plugins: [svelte({ configFile: "./svelte.config.mjs" })],
   test: {
-    projects: ["tests/*"],
+    name: "renderer",
+    environment: "jsdom",
+    include: ["src/**/*.test.ts"],
     coverage: {
       include: ["src/**/*.{ts,svelte}"],
     },
